+++ test/forms/mappings/ConstraintsSpec.scala
@@ -16,7 +16,7 @@
 
 package forms.mappings
 
-import generators.Generators
+import helpers.generators.Generators
 import org.scalacheck.Gen
 import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.must.Matchers
