+++ test/views/btn/BTNSubmitUKTRViewSpec.scala
@@ -16,9 +16,9 @@
 
 package views.btn
 
+import base.ViewSpecBase
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Document
-import utils.ViewSpecBase
 import views.html.btn.BTNSubmitUKTRView
 
 class BTNSubmitUKTRViewSpec extends ViewSpecBase {
