+++ test/views/uktr/UKTaxReturnStartViewSpec.scala
@@ -16,9 +16,9 @@
 
 package views.uktr
 
+import base.ViewSpecBase
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Document
-import utils.ViewSpecBase
 import views.html.uktr.UKTaxReturnStartView
 
 class UKTaxReturnStartViewSpec extends ViewSpecBase {
